decorate the request rather than the reply
