+++ src/apiProxy/apiProxyHandler.js
@@ -1,6 +1,4 @@
-import url from 'url';
 import Boom from 'boom';
-import rison from 'rison';
 
 import 'rxjs/add/operator/catch';
 
@@ -11,15 +9,7 @@ const handleQueryResponses = (request, reply) => queryResponses => {
 		})
 	).type('application/json');
 
-	const payload = request.method === 'post' ? request.payload : request.query;
-
-	const metadataRison = payload.metadata || rison.encode_object({});
-	const metadata = rison.decode_object(metadataRison);
-	const originUrl = response.request.info.referrer;
-	metadata.url = url.parse(originUrl).pathname;
-	metadata.method = response.request.method;
-
-	reply.trackApi(response, queryResponses, metadata);
+	request.trackApi(response, queryResponses);
 };
 
 export const getApiProxyRouteHandler = proxyApiRequest$ => (request, reply) => {
