+++ src/routes/appRouteHandler.js
@@ -22,6 +22,6 @@ export const getAppRouteHandler = renderRequestMap => (request, reply) => {
 			// response is sent when this function returns (`nextTick`)
 			const response = reply(result).code(statusCode);
 
-			reply.trackSession(response);
+			request.trackSession(response);
 		});
 };
