+++ src/util/testUtils.js
@@ -62,8 +62,8 @@ export const getServer = () => {
 		request => () => Observable.of(request),
 		{ apply: true }
 	);
-	server.decorate('reply', 'trackApi', () => ({}));
-	server.decorate('reply', 'trackSession', () => ({}));
+	server.decorate('request', 'trackApi', () => ({}));
+	server.decorate('request', 'trackSession', () => ({}));
 	server.logger = () => MOCK_LOGGER;
 	return server;
 };
